+++ import_export_extensions/models/export_job.py
@@ -11,6 +11,7 @@ from django.utils.translation import gettext_lazy as _
 from celery import current_app, result, states
 from import_export.formats import base_formats
 
+from .. import signals
 from . import tools
 from .core import BaseJob, TaskStateInfo
 
@@ -208,15 +209,10 @@ class ExportJob(BaseJob):
                 ],
             )
         except Exception as error:
-            self.traceback = traceback.format_exc()
-            self.error_message = str(error)[:512]
-            self.export_status = self.ExportStatus.EXPORT_ERROR
-            self.save(
-                update_fields=[
-                    "export_status",
-                    "traceback",
-                    "error_message",
-                ],
+            self._handle_error(
+                error_message=str(error),
+                traceback=traceback.format_exc(),
+                exception=error,
             )
 
     def cancel_export(self) -> None:
@@ -276,10 +272,26 @@ class ExportJob(BaseJob):
                 info=async_result.info,
             )
 
-        # Update job's status in case of exception
+        self._handle_error(
+            error_message=str(async_result.info),
+            traceback=str(async_result.traceback),
+        )
+        return dict(
+            state=async_result.state,
+            info={},
+        )
+
+    def _handle_error(
+        self,
+        error_message: str,
+        traceback: str,
+        exception: Exception | None = None,
+    ):
+        """Update job's status in case of error."""
         self.export_status = self.ExportStatus.EXPORT_ERROR
-        self.error_message = str(async_result.info)[:128]
-        self.traceback = str(async_result.traceback)
+        error_message_limit = self._meta.get_field("error_message").max_length
+        self.error_message = error_message[:error_message_limit]
+        self.traceback = traceback
         self.save(
             update_fields=[
                 "error_message",
@@ -287,7 +299,10 @@ class ExportJob(BaseJob):
                 "export_status",
             ],
         )
-        return dict(
-            state=async_result.state,
-            info={},
+        signals.export_job_failed.send(
+            sender=self.__class__,
+            instance=self,
+            error_message=self.error_message,
+            traceback=self.traceback,
+            exception=exception,
         )