+++ import_export_extensions/signals.py
@@ -1,0 +1,4 @@
+from django import dispatch
+
+import_job_failed = dispatch.Signal()
+export_job_failed = dispatch.Signal()