+++ test_project/fake_app/apps.py
@@ -6,3 +6,7 @@ class IOExtensionsAppConfig(AppConfig):
 
     name = "test_project.fake_app"
     verbose_name = "Import Export Fake App"
+
+    def ready(self):
+        # Import to connect signals.
+        from . import signals  # noqa: F401