+++ import_export_extensions/models/export_job.py
@@ -81,6 +81,7 @@ class ExportJob(BaseJob):
 
     data_file = models.FileField(
         max_length=512,
+        storage=tools.select_storage,
         verbose_name=_("Data file"),
         upload_to=tools.upload_export_file_to,
         help_text=_("File that contain exported data"),