+++ HISTORY.rst
@@ -6,8 +6,8 @@ Unreleased
 ------------------
 
 * Fix issue with long `error_message`
-* Add ability to specify hook to additional handling of import/export errors.
-  Use `IMPORT_EXPORT_JOB_ERROR_HOOK_PATH` in settings.
+* Add signals `export_job_failed` and `import_job_failed`
+  to respond to failed jobs.
 
 1.4.1 (2025-02-18)
 ------------------