+++ HISTORY.rst
@@ -8,6 +8,8 @@ Unreleased
 * Fix issue with long `error_message`
 * Add signals `export_job_failed` and `import_job_failed`
   to respond to failed jobs.
+* Add ability to specify storage via `STORAGES` setting and alias
+  `django_import_export_extensions`
 
 1.4.1 (2025-02-18)
 ------------------